Update to v 1.3 Refactor Files Add missing documentation

+++ tests/tests/test_link.cpp
@@ -52,6 +52,6 @@ void testLink() {
 
   g.waitForTermination();
 
-  ASSERT_EQ(count, 100);
+  ASSERT_EQ(count, (size_t)100);
 #endif
 }
